replace dPLI with dpLI corrected and na_dPLI with na_dPLI corrected
Liz Teel found that some values in the dPLI were 0, which should not be possible. She found that the real values were not filled in the matrix because there are some special cases where the surrogate or dPLI is exactly 0.5. These cases were not covered in the if cases. Therefore 5 new cases were added.